Round rather than truncate the reconstructed Monkhorst-Pack grid
cell_find_MP_grid recovers the grid density as 1/min_img, where min_img is the
smallest separation in the time-reversal-completed, folded k-point list. The
k-points come from the .bands file, which CASTEP writes to 8 decimals, so
min_img is a difference of rounded numbers and the reciprocal misses the
integer by a few ulp in either direction. int() truncates, and a miss on the
low side then costs a whole grid point.

1/17 = 0.0588235294... is written as 0.05882353, i.e. rounded up. Every
pairwise difference inherits that, so

    min_img       = 0.05882352999999996
    1.0_dp/min_img = 16.999999830000014

and a 17x17x1 mesh was reported as 16x16x1.

Which sizes are affected depends on where the decimal rounding happens to land
and is not monotonic in the write precision. Driving the real subroutine over
n = 2..300: at 8 dp (what CASTEP writes) n=17 is the only one; at 9 dp it is
n = 7, 11, 13, 19, 77, 91, 133, 143, 209, 247; at 10 dp, n=101. With nint,
zero at every precision from 7 dp up. So this is not a hazard a user can
avoid by knowing which grid sizes to steer clear of -- the set is an artefact
of the file format, and a CASTEP that wrote one more digit would change it.

Verified end to end on a real calculation rather than a model: Cu(100), one
layer, kpoint_mp_grid 17 17 1, which CASTEP reduces to 45 irreducible
k-points. Two binaries differing only in these two words, same .bands, same
.odi -- the .odo files differ in the "Grid size" line and nothing else.

    -out.cell truth   17  17   1
    int()             16 x 16 x  1
    nint()            17 x 17 x  1

No benchmark moves. The full suite reports the same 54 of 69 and all 1465
compared values are identical either way, because no fixture uses a grid where
the two differ -- which is also why this survived. The smallest fixture that
would catch it is the 17x17x1 case above.

The neighbouring WARNING is extended while here. The comment named the 3/(4n)
shift; the family is wider. With d = (2*shift) mod 1/n and a = min(d, 1/n - d),
a shift aliases onto 2n when a = 1/(2n) and onto 3n when a = 1/(3n), and is
reported correctly otherwise -- so no shift and the half-step shift, the two a
user is normally told to use, are both safe, as is an arbitrary shift. It is
the tidy quarter and third steps that alias. Measured over n = 2..12 and
shifts p/(qn), q = 2..6: 66 mis-reported combinations, unchanged by this
commit. That aliasing is a separate bug and still not fixable here, for the
reason the original comment gives: without the symmetry operations the shifted
mesh cannot be built to rule it out.

Downstream, photo.f90 forms total_ks = kpoint_grid_dim(1)*kpoint_grid_dim(2),
so a mis-read grid makes m_k = kpoint_weight(N_k)*total_ks non-integral and
photo_check_star reports the star bookkeeping as failed for every k-point.

+++ optados/src/cell.f90
@@ -302,17 +302,36 @@ contains
         write (stdout, *) " 1st^-1:", 1.0_dp/min_img
       end if
 
+      ! nint, not int. The k-points are read from the .bands file, where CASTEP
+      ! writes them to 8 decimals, so min_img is a difference of rounded numbers
+      ! and 1/min_img misses the integer by a few ulp either way. Truncating
+      ! then loses a whole grid point: 1/17 rounds UP to 0.05882353, every
+      ! pairwise difference inherits that, and 1.0_dp/min_img comes out
+      ! 16.999999830000014 -- so a 17x17x1 mesh was reported as 16x16x1.
+      ! Which sizes are hit depends on where the decimal rounding lands and is
+      ! not monotonic in the precision: at 8 dp only n=17 in 2..300, at 9 dp
+      ! n = 7, 11, 13, 19, 77, 91, 133, 143, 209, 247. With nint, none at any
+      ! precision from 7 dp up. Verified against a real CASTEP 17x17x1 run.
       if (abs(2.0_dp*min_img - min_img2) < min_img_tol) then
         ! If 1stMI==2ndMI then 1/1stMP is the grid density
         if (present(kpoint_offset)) kpoint_offset(idim) = 0.0_dp
-        kpoint_grid_dim(idim) = int(1.0_dp/min_img)
-        ! WARNING could also have a shifted grid with a perfect shift 3/(4n)
-        ! this would be a known bug
+        kpoint_grid_dim(idim) = nint(1.0_dp/min_img)
+        ! WARNING a shifted grid can still alias onto a denser one, and the
+        ! family is wider than the 3/(4n) named above. With d = (2*shift) mod
+        ! 1/n and a = min(d, 1/n - d), for n >= 2:
+        !     d = 0        (shift = m/(2n))          reported correctly
+        !     a = 1/(2n)   (shift = odd/(4n))        reported as 2n
+        !     a = 1/(3n)   (shift = m/(6n), 3 !| m)  reported as 3n
+        !     otherwise                              reported correctly
+        ! So no shift and the half-step shift -- the two a user is told to use
+        ! -- are both safe, and an arbitrary shift is safe; it is the tidy
+        ! quarter and third steps that alias. Still unfixable here: without the
+        ! symmetry operations we cannot build the shifted mesh to rule it out.
       else
         ! If 1stMI.ne.2ndMI then 1/3rdMP is the grid density
         ! and 1stMI/2 is the shift
         if (present(kpoint_offset)) kpoint_offset(idim) = min_img/2.0_dp
-        kpoint_grid_dim(idim) = int(1.0_dp/min_img3)
+        kpoint_grid_dim(idim) = nint(1.0_dp/min_img3)
       end if
 
     end do over_dim
